morph and blend now use the same approximation this has to be calculated only once

+++ src/preprocess/shaders/split.wgsl
@@ -23,7 +23,7 @@ override fn pixel_value(coords: vec2<u32>) -> vec4<f32> {
     let node_coordinate = split_data.node.coordinate;
     let node_offset =  vec2<f32>(f32(node_coordinate.x), f32(node_coordinate.y));
     let node_coords = vec2<f32>(coords - vec2<u32>(attachment.border_size)) / f32(attachment.center_size);
-    let node_scale = f32(node_count(node_coordinate.lod));
+    let node_scale = node_count(node_coordinate.lod);
 
     var tile_coords = (node_offset + node_coords) / node_scale;
